backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
sqlite3  # Built-in

# Data & APIs
requests>=2.31.0
pyyaml>=6.0.1
pytrends>=4.9.0
praw>=7.7.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional (for production)
gunicorn>=21.2.0
